+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release 2.0.0
 
+* Implemented the HTTPS server using the Twisted framework
 * HEAD requests are now logged too
 * Implemented command-line options
 * Implemented config file